requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-dotenv
pydantic

# Dependencias para Machine Learning y Análisis
scikit-learn
jupyter
matplotlib
seaborn
SQLAlchemy
nltk
xgboost
shap

# Dependencias para Testing
pytest
httpx
